Import implementation from the SDK v1.0.0rc601
The only difference with the SDK is the files are split to have one per
quantity type, to have more manageable files.

Also `_NoDefaultConstructible` was renamed to `NoDefaultConstructible`,
as it is used outside the module it is defined in (`_quantity.py`). It
is not exported in `__init__.py` though. In the future it should
probably be moved to some core library.

The tests are also imported from the SDK but kept in one file, as there
are many tests that use multiple quantity types to compose them.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>